back-end/requirements.txt
beautifulsoup4==4.8.0
certifi==2019.9.11
chardet==3.0.4
Click==7.0
Flask==1.1.1
Flask-SQLAlchemy==2.4.1
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
psycopg2==2.8.4
requests==2.22.0
soupsieve==1.9.4
urllib3==1.25.6
Werkzeug==0.16.0

flask-marshmallow==0.10.1
marshmallow==3.2.1
marshmallow-sqlalchemy==0.19.0


